+++ app/models/survey/Survey.java
@@ -11,14 +11,70 @@ package models.survey;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 
-public class Survey {
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Lob;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+
+import play.data.validation.MaxSize;
+import play.data.validation.MinSize;
+import play.data.validation.Required;
+import play.data.validation.Unique;
+import models.Account;
+import models.ModelWithTimestamp;
+
+@Entity
+@Table(name="survey")
+public class Survey extends ModelWithTimestamp {
   
+  @Required
+  @MinSize(1)
   public String title;
   
-  public List<Question> questions;
+  @Required
+  @ManyToOne
+  public Account createdBy;
+
+  @Required
+  @MinSize(1)
+  @MaxSize(10000)
+  @Lob
+  @Column(columnDefinition="TEXT")
+  public String description;
+
+  @Required
+  @Unique
+  public String code;
+  
+  public Boolean isPublished;
+  
+  //TODO public List<Question> questions;
   
   public Survey() {
-    questions = new ArrayList<Question>();
+    //TODO questions = new ArrayList<Question>();
+    code = generateCode(10);
+    //TODO set isPublished to false, and add an action to publish it manually.
+    isPublished = true;
+    isDeleted = false;
+  }
+
+  
+  private String generateCode(int size) {
+    String chars = "234567890abcdefghijkmnopqrstuvwxyz";
+    String result = "";
+    //TODO prevent duplicate
+    Random gen = new Random();
+    for (int i = 0; i<size; i++)
+      result += chars.charAt(gen.nextInt(chars.length()));
+    return result;
+  }
+  
+  public static Survey findByCode(String code) {
+    return Survey.find("byCode", code).first();
   }
 }